pyproject.toml
[tool.poetry]
name = "py -m poetry add flask sqlalchemy flask-restful flask-cors"
version = "0.1.0"
description = ""
authors = ["Andry2127 <andrybudnak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask-restful = "^0.3.10"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
